refactor: rpctracker -> purple API name
