+++ .github/workflows/build.yml
@@ -42,20 +42,20 @@ jobs:
       - name: Generate Purple API client
         run: |
           container_id="$(docker create ghcr.io/ietf-tools/rpc-purple-backend:latest)"
-          docker cp "${container_id}:/workspace/purple-api.yaml" ./
+          docker cp "${container_id}:/workspace/purple_api.yaml" ./
           docker rm "${container_id}"
           npx --yes @openapitools/openapi-generator-cli generate --generator-key purple
-          tar cfz rpctracker_client.tgz -C client rpctracker_client
+          tar cfz purple_client.tgz -C client purple_client
 
       - name: Upload API
         uses: actions/upload-artifact@v4
         with:
           name: api-clients
           path: |
-            purple-api.yaml
+            purple_api.yaml
             rpcapi.yaml
+            purple_client.tgz
             rpcapi_client.tgz
-            rpctracker_client.tgz
 
       - name: Build frontend image
         uses: docker/build-push-action@v6
