+++ client/components/DocExceptionsCard.vue
@@ -14,7 +14,7 @@
 </template>
 
 <script setup lang="ts">
-import type { Label } from '~/rpctracker_client'
+import type { Label } from '~/purple_client'
 
 type Props = { labels: Label[] }
