+++ openapitools.json
@@ -13,12 +13,12 @@
           "disallowAdditionalPropertiesIfNotPresent": false
         }
       },
-      "rpctracker": {
+      "purple": {
         "generatorName": "typescript-fetch",
-        "inputSpec": "#{cwd}/rpctracker_api.yaml",
-        "output": "#{cwd}/client/rpctracker_client",
+        "inputSpec": "#{cwd}/purple_api.yaml",
+        "output": "#{cwd}/client/purple_client",
         "additionalProperties": {
-          "packageName": "rpctracker_client",
+          "packageName": "purple_client",
           "disallowAdditionalPropertiesIfNotPresent": false
         }
       }
