+++ rpctracker/openapi.py
@@ -5,4 +5,4 @@ class RpcAutoSchema(AutoSchema):
     # pass
     def get_tags(self):
         # Override the default so that we get a single API tag
-        return ["rpc_tracker"]
+        return ["purple"]
